refactor(friends): extract friend search logic into useFriendSearch hook
